Hide select element on iOS devices too
Rather than using width: 0, height: 0, and text-indent to try to hide the select element, simply make it display: none. This is not only shorter and more descriptive, but also fixes a bug where the actual select element would still be visible on iOS devices. Does not appear to cause regressions on IE8/9, Chrome, Safari, or Firefox.