Freeform Updates
Added method, via a custom "data-value" attribute on the select element of the combo-box, to enable the passing of freeform data back to the combo-box control for later display on refresh of a form. Also added ability to use the down-arrow while on a combo-box to open the drop-down option list.